requirements.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
filelock==3.9.0
Flask==2.2.3
gdown==4.6.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pickleDB==0.9.2
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
tqdm==4.64.1
urllib3==1.26.14
Werkzeug==2.2.3
